alterações

+++ acutis_new/src/acutis_api/application/use_cases/agape/deletar/excluir_item_estoque_agape.py
@@ -1,4 +1,5 @@
 from acutis_api.domain.repositories.agape import AgapeRepositoryInterface
+from acutis_api.exception.errors.not_found import HttpNotFoundError
 
 
 class ExcluirItemEstoqueAgapeUseCase:
@@ -18,7 +19,8 @@ class ExcluirItemEstoqueAgapeUseCase:
     ) -> None:
         # Busca o item de estoque pelo ID
         item = self.__repository.buscar_item_estoque_por_id(item_id)
-
+        if item is None:
+            raise HttpNotFoundError('Item não encontrado.')
         # Remove o item do banco
         self.__repository.remover_item_estoque(item)
         self.__repository.salvar_alteracoes()
